% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics.R
\docType{methods}
\name{comp_categ_counts}
\alias{comp_categ_counts}
\alias{comp_categ_counts,hdpSampleChain-method}
\alias{comp_categ_counts,hdpSampleMulti-method}
\title{Get sample vs category counts for each component}
\usage{
comp_categ_counts(x)

\S4method{comp_categ_counts}{hdpSampleChain}(x)

\S4method{comp_categ_counts}{hdpSampleMulti}(x)
}
\arguments{
\item{x}{hdpSampleChain or hdpSampleMulti}
}
\value{
List of matrices (one for each component)
 counting the sample-category data assignment across all DP nodes.
 Number of rows is the number of posterior samples, and number of
 columns is the number of data categories.
}
\description{
Get sample vs category counts for each component
}
\section{Methods (by class)}{
\itemize{
\item \code{hdpSampleChain}: For hdpSampleChain

\item \code{hdpSampleMulti}: For hdpSampleMulti
}}

