% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics-output.R
\name{comp_cos_merge}
\alias{comp_cos_merge}
\title{Get cos.merge setting}
\usage{
comp_cos_merge(x)
}
\arguments{
\item{x}{hdpSampleChain or hdpSampleMulti}
}
\value{
number of components
}
\description{
Get cos.merge setting
}

