% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics-output.R
\name{comp_dp_distn}
\alias{comp_dp_distn}
\title{Get mean distribution over components for each DP}
\usage{
comp_dp_distn(x)
}
\arguments{
\item{x}{hdpSampleChain or hdpSampleMulti}
}
\value{
List with elements "mean" and "cred.int", containing
 matrices with the mean (and lower/upper 95\% credibility interval) distribution
 over components for each DP. Number of rows is the number of
 DPs, and number of columns is the number of components. Rows sum to 1.
}
\description{
Get mean distribution over components for each DP
}
