% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaa-generics.R
\docType{methods}
\name{comp_dp_weights}
\alias{comp_dp_weights}
\alias{comp_dp_weights,hdpSampleChain-method}
\alias{comp_dp_weights,hdpSampleMulti-method}
\title{Get sample vs component weights for each DP}
\usage{
comp_dp_weights(x)

\S4method{comp_dp_weights}{hdpSampleChain}(x)

\S4method{comp_dp_weights}{hdpSampleMulti}(x)
}
\arguments{
\item{x}{hdpSampleChain or hdpSampleMulti}
}
\value{
List of matrices (one for each DP)
 with the weights of each component per sample. Number of rows is the number of
 posterior samples, and number of columns is the number of components.
}
\description{
Get sample vs component weights for each DP
}
\section{Methods (by class)}{
\itemize{
\item \code{hdpSampleChain}: 

\item \code{hdpSampleMulti}: 
}}

