% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cull_posterior_samples.R
\name{cull_posterior_samples}
\alias{cull_posterior_samples}
\title{Cull early posterior samples from a hdpSampleChain object}
\usage{
cull_posterior_samples(chain, ncull)
}
\arguments{
\item{chain}{A hdpSampleChain object}

\item{ncull}{The number of posterior samples to cull}
}
\value{
A hdpSampleChain object with the designated 'burn-in' period extended,
 and the number of posterior samples reduced by \code{ncull}
}
\description{
Extend the 'burn-in' period and reduce the number of posterior samples taken
from a sampling chain by culling the first \code{ncull} posterior samples.
If components have been previously calculated for this sampling chain, they
will be removed and must be recalculated via \code{\link{hdp_extract_components}}.
}
\examples{
chain_adj <- cull_posterior_samples(tcga_example_chain, 20)
plot_lik(chain_adj)
plot_numcluster(chain_adj)
plot_data_assigned(chain_adj)
}
\seealso{
\code{\link{plot_lik}}, \code{\link{plot_numcluster}},
 \code{\link{plot_data_assigned}}
}

