% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic_in_extraction.R
\name{diagnostic_in_extraction}
\alias{diagnostic_in_extraction}
\title{Diagnostic plotting inside of hdp_merge_and_extract_components function.
This function generates details of the raw clusters in hdp.0}
\usage{
diagnostic_in_extraction(
  clust_hdp0_ccc,
  ncat,
  nsamp,
  nch,
  ccc,
  cdc,
  diagnostic.folder
)
}
\arguments{
\item{clust_hdp0_ccc}{An object in \code{\link{hdp_merge_and_extract_components}}.}

\item{ncat}{Number of categories. An internal object from \code{\link{hdp_merge_and_extract_components}}.}

\item{nsamp}{Number of posterior samples. An internal object from \code{\link{hdp_merge_and_extract_components}}.}

\item{nch}{Number of posterior chains. An internal object from \code{\link{hdp_merge_and_extract_components}}.}

\item{ccc}{clust_categ_counts from \code{\link{hdp_merge_and_extract_components}}.}

\item{cdc}{clust_dp_counts from \code{\link{hdp_merge_and_extract_components}}.}

\item{diagnostic.folder}{A directory where details for hdp.0 are plotted.}
}
\value{
The plots of presence of a raw cluster in each chain.
}
\description{
Diagnostic plotting inside of hdp_merge_and_extract_components function.
This function generates details of the raw clusters in hdp.0
}
\seealso{
\code{\link{hdp_merge_and_extract_components}}
}
