% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dp_activate}
\alias{dp_activate}
\title{Activates the DPs to be included in the posterior sampling process.}
\usage{
dp_activate(hdp, dpindex, initcc)
}
\arguments{
\item{hdp}{A HDP list object.}

\item{dpindex}{A vector of indices indicating which DPs in the HDP object to activate.}

\item{initcc}{A positive integer specifying how many clusters to start with.}
}
\description{
Activates the DPs to be included in the posterior sampling process.
}

