% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dp_freeze}
\alias{dp_freeze}
\title{Freezes previously active DPs. A frozen DP is not included in posterior sampling, but the statistics for this DP are used in the sampling of other active DPs.}
\usage{
dp_freeze(hdp, dpindex)
}
\arguments{
\item{hdp}{A HDP list object.}

\item{dpindex}{A vector of indices indicating which DPs in the HDP object to freeze.}
}
\description{
Freezes previously active DPs. A frozen DP is not included in posterior sampling, but the statistics for this DP are used in the sampling of other active DPs.
}

