% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp.R
\docType{data}
\name{example_data_hdp}
\alias{example_data_hdp}
\title{Fake categorical count data}
\format{A numeric count matrix with 10 rows and 6 columns}
\usage{
example_data_hdp
}
\description{
Fake categorical count data with 10 samples and 6 categories.
Generated from two underlying categorical data distributions with a
different average mixture ratio in the first five samples from the last
five samples.
}
\keyword{datasets}
