% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp.R
\docType{data}
\name{example_data_hdp_prior}
\alias{example_data_hdp_prior}
\title{Fake categorical count data with priors}
\format{A numeric count matrix with 100 rows and 10 columns}
\usage{
example_data_hdp_prior
}
\description{
Fake categorical count data with 100 samples and 10 categories.
Generated from four underlying categorical data distributions.
Two of the underlying components are available as known prios
in \code{\link{example_known_priors}}.
}
\keyword{datasets}

