% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp.R
\docType{data}
\name{example_known_priors}
\alias{example_known_priors}
\title{Example known priors}
\format{A numeric matrix with 10 rows and 2 columns.}
\usage{
example_known_priors
}
\description{
Two example prior components for the example data
\code{\link{example_data_hdp_prior}}. 10 rows (one per data category) and
2 columns (one per prior component, each sums to 1).
}
\keyword{datasets}
