% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_plot_functions.R
\name{extract_ccc_cdc_from_hdp}
\alias{extract_ccc_cdc_from_hdp}
\title{find the ccc and cdc that matched to a spectrum in ccc_0 and cdc_0
this function is to summarize the credint and mean of cccs and cdcs
and for further diagnostic plotting}
\usage{
extract_ccc_cdc_from_hdp(spectrum, ccc_0, cdc_0, cos.merge = 0.9)
}
\arguments{
\item{spectrum}{signature spectrum for compare}

\item{ccc_0}{a list object contains clust_categ_counts matrix from hdp}

\item{cdc_0}{a list object contains clust_dp_counts matrix from hdp}

\item{cos.merge}{cosine similarity cutoff}
}
\description{
find the ccc and cdc that matched to a spectrum in ccc_0 and cdc_0
this function is to summarize the credint and mean of cccs and cdcs
and for further diagnostic plotting
}
