% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_components_from_clusters.R
\name{extract_components_from_clusters}
\alias{extract_components_from_clusters}
\title{Combine "raw clusters" of mutations in "components" (clusters of mutations generated by one mutational process).}
\usage{
extract_components_from_clusters(x, cos.merge = 0.9, hc.cutoff = 0.12)
}
\arguments{
\item{x}{An \code{\link{hdpSampleChain-class}} or \
code{\link{hdpSampleMulti-class}} object}

\item{cos.merge}{Merge raw cluster that have cosine similarity above this
threshold.}

\item{hc.cutoff}{the height to cut hierarchical clustering tree}
}
\value{
A list with the elements \describe{
\item{components}{Clusters profile as a data frame.Rows represent the categories and columns are index.
               each cell contains number of items}
\item{components.post.samples}{A data frame with two columns: one is the index and the other is number of posterior
                             samples that extract raw clusters contributing to \code{components}}
\item{components.cdc}{A categ_dp_counts matrix. Each row is a dp and each column corresponds to the cluster
                    in \code{components}}
\item{each.chain.noise.clusters}{A list of raw clusters that only found in one posterior sample of each chain.
                             These clusters were selected before hierarchical clustering to save computational time}
\item{each.chain.noise.cdc}{A categ_dp_counts matrix with each row is a dp and each column corresponds to
                           the cluster in \code{each.chain.noise.clusters}}
\item{multi.chains}{An \code{\link{hdpSampleChain-class}}
   or \code{\link{hdpSampleMulti-class}}
   object updated with component information
}
\item{nsamp}{The total number of posterior samples.}
}
}
\description{
Combine "raw clusters" of mutations in "components" (clusters of mutations generated by one mutational process).
}
\seealso{
\code{\link{hdp_posterior}}, \code{\link{hdp_multi_chain}},
 \code{\link{plot_comp_size}}, \code{\link{plot_comp_distn}},
 \code{\link{plot_dp_comp_exposure}}
}
