% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sigs_from_clusters.R
\name{extract_sigs_from_clusters}
\alias{extract_sigs_from_clusters}
\title{This function combines part of NR's code and new way to look at sigantures from raw clusters
This is a testing function that maybe or not work}
\usage{
extract_sigs_from_clusters(x, cos.merge, prop.samp, min.sample)
}
\arguments{
\item{x}{hdpSampleChain or hdpSampleMulti object}

\item{cos.merge}{Merge components with cosine similarity above this threshold (default 0.90)}

\item{prop.samp}{The proportion of posterior samples with a certain cluster to be determined as                     a signature. e.g., if prop.samp = 0.5, a cluster will be determined as a                           signature if it appears in half of the posterior samples}

\item{min.sample}{Components must have significant exposure in at least this many samples (i.e. those DP nodes with data assigned) (default 1)}
}
\value{
A hdpSampleChain or hdpSampleMulti object updated with component information
}
\description{
This function combines part of NR's code and new way to look at sigantures from raw clusters
This is a testing function that maybe or not work
}
\seealso{
\code{\link{hdp_posterior}}, \code{\link{hdp_multi_chain}},
 \code{\link{plot_comp_size}}, \code{\link{plot_comp_distn}},
 \code{\link{plot_dp_comp_exposure}}
}
