% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_utils.R
\name{generate_label_high_cossim}
\alias{generate_label_high_cossim}
\title{label_based_on_cosine_similarity}
\usage{
generate_label_high_cossim(clust_label, matrix, cos.sim)
}
\arguments{
\item{clust_label}{column names of the input matrix}

\item{matrix}{the matrix needs to be measured}

\item{cos.sim}{cosine similarity threshold}
}
\description{
label_based_on_cosine_similarity
}
\keyword{internal}
