% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_addconparam.R
\name{hdp_addconparam}
\alias{hdp_addconparam}
\title{Add concentration parameters to a hdpState object}
\usage{
hdp_addconparam(hdp, alphaa, alphab)
}
\arguments{
\item{hdp}{A hdpState object}

\item{alphaa}{Shape hyperparameters for the gamma priors over the DP concentration parameters.}

\item{alphab}{Rate hyperparameters for the gamma priors over the DP concentration parameters.}
}
\value{
A hdpState object updated with the new concentration parameters. See \code{\link{hdpState-class}}
}
\description{
Add concentration parameters to a hdpState object by specifying the shape and
rate parameters of the gamma prior/s. DPs using these new concentration parameters
can be added with \code{\link{hdp_adddp}}. Data is assigned via \code{\link{hdp_setdata}}.
When initialised, the DP nodes are 'heldout' (not available for posterior sampling)
and will need to be activated (see \code{\link{dp_activate}}). Finally, the posterior
sampling process (a Gibbs sampler) is run via \code{\link{hdp_posterior}}.
}
\examples{
hdp_example <- hdp_init(c(0, 1, 1), c(1, 2, 2), rep(1, 6), rep(2, 2), rep(0.5, 2))
hdp_example <- hdp_addconparam(hdp_example, rep(1, 2), rep(1, 2))
}
\seealso{
\code{\link{hdp_init}}, \code{\link{hdp_adddp}}, \code{\link{hdp_setdata}},
 \code{\link{dp_activate}}, \code{\link{hdp_posterior}}
}
