% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hdp_adddp}
\alias{hdp_adddp}
\title{Add DPs to the HDP list object.}
\usage{
hdp_adddp(hdp, numdp, pp, cp)
}
\arguments{
\item{hdp}{A HDP list object.}

\item{numdp}{The number of DPs to add (a positive integer).}

\item{pp}{Integer specifying the index of the parental process for the new DPs.}

\item{cp}{Integer specifying the index of the concentration parameters for the new DPs.}
}
\value{
A list with the updated HDP structure.
}
\description{
Add DPs to the HDP list object.
}

