% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_burnin.R
\name{hdp_burnin}
\alias{hdp_burnin}
\title{Burnin of posterior sampling chain across activated DPs.}
\usage{
hdp_burnin(hdp, burnin, cpiter = 1, verbosity = 0)
}
\arguments{
\item{hdp}{An \code{\link[hdpx]{hdpState-class}} object.}

\item{burnin}{The number of burn-in iterations.}

\item{cpiter}{The number of iterations of concentration parameter sampling
to perform after each iteration.}

\item{verbosity}{Verbosity of debugging statements.
0 (least verbose) -- 4 (most verbose). 0 highly recommended
 - only change for debugging small examples.}
}
\value{
A list with the elements: \describe{

\item{hdplist}{A list representation of
   an \code{\link[hdpx]{hdpState-class}} object.}

\item{likelihood}{A numeric vector with the likelihood at each iteration.}

}
}
\description{
Run a Gibbs sampler over the activated nodes of a hierarchical Dirichlet process.
Each iteration re-assigns the cluster allocation of every data item.
}
\examples{
my_hdp <- hdp_init(ppindex=0, cpindex=1, hh=rep(1, 6), alphaa=rep(1, 3), alphab=rep(2, 3))
my_hdp <- hdp_adddp(my_hdp, 2, 1, 2)
my_hdp <- hdp_adddp(my_hdp, 10, c(rep(2, 5), rep(3, 5)), 3)
my_hdp <- hdp_setdata(my_hdp, 4:13, example_data_hdp)
my_hdp <- dp_activate(my_hdp, 1:13, 2)
my_hdp_chain <- hdp_burnin(my_hdp, burnin = 100, cpiter = 100)
}
