% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdp_extract_components.R
\name{hdp_extract_components}
\alias{hdp_extract_components}
\title{Extract major components from the raw clusters of a hdpSampleChain}
\usage{
hdp_extract_components(chain, prop.ex = 0.97, cos.merge = 0.9)
}
\arguments{
\item{chain}{A hdpSampleChain object}

\item{prop.ex}{The proportion of data items explained by the components (default 0.97)}

\item{cos.merge}{Merge components with cosine similarity above this threshold (default 0.90)}
}
\value{
A hdpSampleChain object updated with component information
}
\description{
Extract major components from the raw clusters of a hdpSampleChain
}
\examples{
hdp_extract_components(tcga_example_chain)
}
\seealso{
\code{\link{hdp_posterior}}, \code{\link{plot_comp_size}},
 \code{\link{plot_comp_distn_bar}}, \code{\link{plot_dp_comp_exposure}}
}

