% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hdp_extract_signatures}
\alias{hdp_extract_signatures}
\title{Consolidates the large number of classes into a smaller number of overall signatures.}
\usage{
hdp_extract_signatures(output, prop.explained, cos.merge)
}
\arguments{
\item{output}{The output list object from hdp_posterior}

\item{prop.explained}{The proportion of data points to assign to overall signatures (e.g. 0.95)}

\item{cos.merge}{Merge signatures with cosine similarity greater than this (e.g. 0.90)}
}
\value{
A list with three elements:  map_classes_to_sigs - the mapping from class columns in the original classqq to the signature columns in sigs_qq; sigs_qq - a list of matrices summarising the total number of data points of each category in each overall signature at each posterior sample collected (signature 0 includes all unassigned data points); and sigs_nd_by_dp - a list of matrices summarising the total number of data points assigned to each overall signature in each posterior sample collected, for each DP in the HDP (signature 0 includes all unassigned data points).
}
\description{
Consolidates the large number of classes into a smaller number of overall signatures.
}

