% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hdp_init}
\alias{hdp_init}
\title{Initialises a HDP list object.}
\usage{
hdp_init(ppindex, cpindex, hh, alphaa, alphab)
}
\arguments{
\item{ppindex}{Integer (or vector of integers) specifying the indices of the parental process(es) for the initial DPs. The 'top' DP will have parent process '0'.}

\item{cpindex}{Integer (or vector of integers) specifying the indices of the concentration parameters for the initial DPs.}

\item{hh}{Numeric vector with one entry for each category in the dataset; specifies the parameters of the base Dirichlet distribution. To specify the uniform Dirichlet, set hh=rep(1/number_of_categories, number_of_categories).}

\item{alphaa}{Numeric vector of shape hyperparameters for the gamma priors for DP concentration parameters.}

\item{alphab}{Numeric vector of inverse scale hyperparameters for the gamma priors for DP concentration parameters.}
}
\value{
A list with the initial HDP structure.
}
\description{
Initialises a HDP list object.
}

