% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_merge_and_extract_components.R
\name{hdp_merge_and_extract_components}
\alias{hdp_merge_and_extract_components}
\alias{hdp_extract_components}
\title{Extract major components from the raw clusters}
\usage{
hdp_merge_and_extract_components(
  x,
  cluster.method = "kmedians",
  categ.CI = 0.95,
  exposure.CI = 0.95,
  cos.merge = 0.9,
  min.sample = 1,
  diagnostic.folder = NULL
)
}
\arguments{
\item{x}{hdpSampleChain or hdpSampleMulti object}

\item{cluster.method}{a temporary argument+}

\item{categ.CI}{A numeric between 0 and 1. Level of confidence interval to be calculated for each category.
Default is 0.95, but can be set to lower for extracting rare signatures}

\item{exposure.CI}{A numeric between 0 and 1. Level of confidence interval to be calculated
for a sample's exposure/observation of a raw cluster/proto-signature.
Default is 0.95, but can be set to lower for extracting rare signatures}

\item{cos.merge}{Merge components with cosine similarity above this threshold (default 0.90)}

\item{min.sample}{Components must have significant exposure in at least this many samples (i.e. those DP nodes with data assigned) (default 1)}

\item{diagnostic.folder}{If provided, details for hdp.0 is plotted}
}
\value{
A hdpSampleChain or hdpSampleMulti object updated with component information
}
\description{
If prior components included via \code{\link{hdp_prior_init}} are
preserved by \code{hdp_extract_components}, they are prefixed with "P".
Any new components in this case are prefixed with "N".
}
\seealso{
\code{\link{hdp_posterior}}, \code{\link{hdp_multi_chain}},
 \code{\link{plot_comp_size}}, \code{\link{plot_comp_distn}},
 \code{\link{plot_dp_comp_exposure}}
}
