% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_multi_chain.R
\name{hdp_multi_chain}
\alias{hdp_multi_chain}
\title{Gather multiple independent posterior sampling chains from one HDP}
\usage{
hdp_multi_chain(chain_list)
}
\arguments{
\item{chain_list}{A list of hdpSampleChain objects for the same data and HDP structure, but run with different random seeds.}
}
\value{
A hdpSampleMulti object
}
\description{
Gather multiple independent posterior sampling chains from one HDP
}
\seealso{
\code{\link{hdp_posterior}}
}

