% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp.R
\docType{data}
\name{hdp_p}
\alias{hdp_p}
\title{HDP posterior samples on example data with priors}
\format{hdpSampleMulti object}
\usage{
hdp_p
}
\description{
Four independent posterior sampling chains for the dataset
\code{\link{example_data_hdp_prior}}, conditioning on two known prior components
provided in \code{\link{example_known_priors}}.
}
\keyword{datasets}
