% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdp_posterior.R
\name{hdp_posterior}
\alias{hdp_posterior}
\title{Posterior sampling chain across activated DPs.}
\usage{
hdp_posterior(hdp, burnin, n, space, cpiter = 1, seed = sample(1:10^7, 1),
  verbosity = 0)
}
\arguments{
\item{hdp}{A hdpState object}

\item{burnin}{The number of burn-in iterations.}

\item{n}{The number of posterior samples to collect.}

\item{space}{The number of iterations between collected samples.}

\item{cpiter}{The number of iterations of concentration parameter sampling to perform after each iteration.}

\item{seed}{The (integer) seed that can be set to reproduce output. Default is a
random seed from 1 -- 10^7, reported in the output.}

\item{verbosity}{Verbosity of debugging statements.
0 (least verbose) -- 4 (most verbose). 0 highly recommended - only change for debugging small examples.}
}
\value{
A hdpSampleChain object with the salient information from each
 posterior sample. See \code{\link{hdpSampleChain-class}}
}
\description{
Run a Gibbs sampler over the activated DP nodes of a Hierarchichal Dirichlet Process.
Each iteration re-assigns the cluster allocation of every data item.
Run \code{burnin} iterations, and then collect \code{n} samples from the chain
with \code{space} iterations between each collected sample.
}
\examples{
my_hdp <- hdp_init(ppindex=0, cpindex=1, hh=rep(1, 6), alphaa=rep(1, 3), alphab=rep(2, 3))
my_hdp <- hdp_adddp(my_hdp, 2, 1, 2)
my_hdp <- hdp_adddp(my_hdp, 10, c(rep(2, 5), rep(3, 5)), 3)
my_hdp <- hdp_setdata(my_hdp, 4:13, example_data_hdp)
my_hdp <- dp_activate(my_hdp, 1:13, 2)
my_hdp_chain <- hdp_posterior(my_hdp, 100, 100, 10)
}
\seealso{
\code{\link{hdp_quick_init}}, \code{\link{hdp_init}}, \code{\link{hdp_addconparam}},
 \code{\link{hdp_adddp}}, \code{\link{hdp_setdata}}, \code{\link{dp_activate}}, \code{\link{dp_freeze}}
}

