% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_posterior_sample.R
\name{hdp_posterior_sample}
\alias{hdp_posterior_sample}
\title{Posterior sampling chain across activated DPs.}
\usage{
hdp_posterior_sample(
  post.input,
  post.n,
  post.space,
  post.cpiter = 1,
  seed = sample(1:10^7, 1),
  post.verbosity = 0,
  checkpoint = F
)
}
\arguments{
\item{post.input}{An S4 object from \code{\link{hdp_burnin}}.}

\item{post.n}{The number of posterior samples to collect.}

\item{post.space}{The number of iterations between collected samples.}

\item{post.cpiter}{The number of iterations of concentration parameter sampling to perform after each iteration.}

\item{seed}{The (integer) seed that can be set to reproduce output. Default is a
random seed from 1 -- 10^7, reported in the output.}

\item{post.verbosity}{Verbosity of debugging statements.
0 (least verbose) -- 4 (most verbose). 0 highly recommended - only change for debugging small examples.}

\item{checkpoint}{If \code{TRUE}, a checkpoint will be saved for every 10 posterior samples}
}
\value{
A hdpSampleChain object with the salient information from each
 posterior sample. See \code{\link{hdpSampleChain-class}}
}
\description{
Run a Gibbs sampler over the burnin chains from \code{\link{hdp_burnin}}.
Each iteration re-assigns the cluster allocation of every data item.
Run \code{burnin} iterations, and then collect \code{n} samples from the chain
with \code{space} iterations between each collected sample. To plot output,
see \code{\link{plot_lik}}, \code{\link{plot_numcluster}}, and
\code{\link{plot_data_assigned}}. Can collect multiple
independent HDP sampling chains in a hdpSampleMulti object via \code{\link{hdp_multi_chain}}.
Components are extracted via \code{\link{hdp_extract_components}}.
}
\seealso{
\code{\link{hdp_multi_chain}}, \code{\link{hdp_extract_components}},
 \code{\link{cull_posterior_samples}}, \code{\link{plot_lik}}, \code{\link{plot_numcluster}},
 \code{\link{plot_data_assigned}}
}
