% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_prior_init.R
\name{hdp_prior_init}
\alias{hdp_prior_init}
\title{Initialise a HDP structure incorporating prior knowledge}
\usage{
hdp_prior_init(prior_distn, prior_pseudoc, hh, alphaa, alphab)
}
\arguments{
\item{prior_distn}{Matrix of prior distributions (columns must each sum to 1,
number of rows matches number of data categories)}

\item{prior_pseudoc}{Vector of pseudocounts contributed by each prior distribution}

\item{hh}{Parameters of the base Dirichlet distribution. Must be a vector with length equal
to the number of data item categories.}

\item{alphaa}{Shape hyperparameters for the gamma priors over the DP concentration parameters.}

\item{alphab}{Rate hyperparameters for the gamma priors over the DP concentration parameters.}
}
\value{
A hdpState object with one frozen node per prior component. See \code{\link{hdpState-class}}
}
\description{
Initialise a hdpState object incorporating prior knowlegde of some components
(categorical data distributions). The structure has one top parent DP node
with no associated data ('active' and available for posterior sampling),
and one child DP node per prior component ('frozen' and held out
from posterior sampling).
}
\examples{
# example dataset with 10 data categories, and 100 samples.
# Two components are known a priori.
hdp <- hdp_prior_init(example_known_priors, rep(1000, 2), hh=rep(1, 10),
             alphaa=c(1,1), alphab=c(1,1))
hdp <- hdp_addconparam(hdp, alphaa=c(1,1), alphab=c(1,1))
hdp <- hdp_adddp(hdp, 101, c(1, rep(4, 100)), c(3, rep(4, 100)))
hdp <- hdp_setdata(hdp, 5:104, example_data_hdp_prior)
hdp <- dp_activate(hdp, 4:104, initcc=4, seed=81479)
hdp <- hdp_posterior(hdp, burnin=2000, n=50, space=50, cpiter=3, seed=1e6)
hdp_ex <- hdp_extract_components(hdp)
plot_comp_size(hdp_ex)
plot_comp_distn(hdp_ex)
plot_dp_comp_exposure(hdp_ex, 5:104, col_comp=rainbow(5))
}
\seealso{
\code{\link{hdp_init}}
}
