% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_quick_init.R
\name{hdp_quick_init}
\alias{hdp_quick_init}
\title{Initialise a simple, default HDP structure}
\usage{
hdp_quick_init(data, initcc = 2, alphaa = 1, alphab = 1)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} of counts with one row for every sample
and one column for every data category.}

\item{initcc}{Number of initial data clusters
(every data item is randomly assigned to a cluster to start with).}

\item{alphaa}{Shape hyperparameter for the gamma prior over the concentration parameter.}

\item{alphab}{Rate hyperparameter for the gamma prior over the concentration parameter.}
}
\value{
A hdpState object with a basic default structure. See \code{\link{hdpState-class}}
}
\description{
Initialise a hdpState object with a basic default structure of one top parent DP node
with no associated data, and one child DP node per row of data. Every DP node
shares the same concentration parameter, and will automatically be 'activated'
(made available for posterior samplig). The base distribution is a uniform Dirichlet
with psuedocount 1 in each data category. Can immediately run \code{\link{hdp_posterior}}
to collect posterior samples.
To define a custom HDP structure, see \code{\link{hdp_init}} and \code{\link{hdp_prior_init}}.
}
\examples{
my_quick_hdp <- hdp_quick_init(example_data_hdp)
my_quick_hdp_chain <- hdp_posterior(my_quick_hdp, 100, 50, 10, 5)
}
\seealso{
\code{\link{hdp_init}}, \code{\link{hdp_posterior}}, \code{\link{hdp_prior_init}}
}
