% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hdp_setdata}
\alias{hdp_setdata}
\title{Assign data to DPs in a HDP list object.
The DPs having data assigned must still be 'held out'.}
\usage{
hdp_setdata(hdp, dpindex, data)
}
\arguments{
\item{hdp}{A HDP list object.}

\item{dpindex}{The indices of the DPs to assign data to.}

\item{data}{A data.frame with: one row for every sample to assign; one column for every category; counts of that data category in that sample.}
}
\value{
A list with the updated HDP structure.
}
\description{
Assign data to DPs in a HDP list object.
The DPs having data assigned must still be 'held out'.
}

