% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaa-classes.R, R/aaa-generics.R
\docType{class}
\name{hdpBase-class}
\alias{as.list,hdpBase-method}
\alias{hdpBase-class}
\title{hdpBase class for the base distribution}
\usage{
\S4method{as.list}{hdpBase}(x, ...)
}
\arguments{
\item{x}{Object of class hdpBase}

\item{...}{unused}
}
\description{
hdpBase class for the base distribution
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.list}: convert to list class
}}
\section{Slots}{

\describe{
\item{\code{hh}}{parameters for base Dirichlet distribution (pseudocounts)}

\item{\code{classqq}}{overall count matrix for data items of each category (rows) in each cluster (columns)}

\item{\code{numclass}}{number of clusters}
}}

