% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R, R/aaa-generics.R
\docType{class}
\name{hdpConparam-class}
\alias{as.list,hdpConparam-method}
\alias{hdpConparam-class}
\title{hdpConparam class for the DP concentration parameter/s}
\usage{
\S4method{as.list}{hdpConparam}(x, ...)
}
\arguments{
\item{x}{Object of class hdpConparam}

\item{...}{unused}
}
\description{
hdpConparam class for the DP concentration parameter/s
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.list}: convert to list class
}}
\section{Slots}{

\describe{
\item{\code{alphaa}}{shape parameter for the gamma prior over alpha}

\item{\code{alphab}}{rate parameter for the gamma prior over alpha}

\item{\code{numdp}}{number of DPs sharing this concentration parameter}

\item{\code{alpha}}{concentration parameter value}

\item{\code{totalnd}}{number of items in each DP with this concentration parameter}

\item{\code{totalnt}}{number of tables in each DP with this concentration parameter}
}}

