% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes.R, R/aaa-generics.R
\docType{class}
\name{hdpDP-class}
\alias{as.list,hdpDP-method}
\alias{hdpDP-class}
\title{hdpDP class for a DP node}
\usage{
\S4method{as.list}{hdpDP}(x, ...)
}
\arguments{
\item{x}{Object of class hdpDP}

\item{...}{unused}
}
\description{
note that the 'items' in parent nodes are the tables of their children
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.list}: convert to list class
}}
\section{Slots}{

\describe{
\item{\code{datacc}}{cluster index for each data item}

\item{\code{classnd}}{number of items assigned to each cluster in this DP}

\item{\code{classnt}}{number of tables assigned to each cluster in this DP}

\item{\code{beta}}{weight on each cluster in this DP (including empty cluster at end)}

\item{\code{alpha}}{concentration parameter for this DP}

\item{\code{numdata}}{number of data items registered to this DP node}

\item{\code{datass}}{value of each data item}
}}

