% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HDPMarginalSampler}
\alias{HDPMarginalSampler}
\title{HDP Marginal Sampler}
\usage{
HDPMarginalSampler(
  Niter,
  Nburnin,
  d,
  n_j,
  data_list,
  priorMean,
  priorA,
  priorB,
  priorLambda,
  a_gamma,
  b_gamma,
  a_alpha,
  b_alpha
)
}
\description{
HDP Marginal Sampler
}
