% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes-output.R, R/aaa-generics-output.R
\docType{class}
\name{hdpSampleMulti-class}
\alias{as.list,hdpSampleMulti-method}
\alias{chains}
\alias{chains,hdpSampleMulti-method}
\alias{comp_categ_counts,hdpSampleMulti-method}
\alias{comp_categ_distn,hdpSampleMulti-method}
\alias{comp_cos_merge,hdpSampleMulti-method}
\alias{comp_dp_counts,hdpSampleMulti-method}
\alias{comp_dp_distn,hdpSampleMulti-method}
\alias{hdpSampleMulti-class}
\alias{numcomp,hdpSampleMulti-method}
\alias{prop.ex,hdpSampleMulti-method}
\title{hdpSampleMulti class for multiple independent hdpSampleChain objects for the same HDP}
\usage{
\S4method{as.list}{hdpSampleMulti}(x, ...)

\S4method{chains}{hdpSampleMulti}(x, ...)

\S4method{numcomp}{hdpSampleMulti}(x)

\S4method{prop.ex}{hdpSampleMulti}(x)

\S4method{comp_cos_merge}{hdpSampleMulti}(x)

\S4method{comp_categ_counts}{hdpSampleMulti}(x)

\S4method{comp_dp_counts}{hdpSampleMulti}(x)

\S4method{comp_categ_distn}{hdpSampleMulti}(x)

\S4method{comp_dp_distn}{hdpSampleMulti}(x)
}
\arguments{
\item{x}{Object of class hdpSampleMulti}

\item{...}{unused}
}
\description{
hdpSampleMulti class for multiple independent hdpSampleChain objects for the same HDP
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.list}: Convert to list class

\item \code{chains}: Get list of hdpSampleChain objects

\item \code{numcomp}: Get number of extracted components for hdpSampleMulti

\item \code{prop.ex}: Get proportion of dataset explained (on average) for hdpSampleMulti

\item \code{comp_cos_merge}: Get cos.merge setting for hdpSampleMulti

\item \code{comp_categ_counts}: Get sample vs category counts for each component

\item \code{comp_dp_counts}: Get sample vs component counts for each DP

\item \code{comp_categ_distn}: Get mean distribution over data categories for each component

\item \code{comp_dp_distn}: Get mean distribution over components for each DP
}}
\section{Slots}{

\describe{
\item{\code{chains}}{List of hdpSampleChain objects storing multiple independent runs of the posterior sampling chain for the same data and HDP struct}

\item{\code{numcomp}}{Number of global components extracted by \code{hdp_extract_components}
(not including component 0)}

\item{\code{prop.ex}}{(Average) proportion of dataset explained by the extracted components}

\item{\code{comp_cos_merge}}{\code{cos.merge} setting used by \code{hdp_extract_components}}

\item{\code{comp_categ_counts}}{List of matrices (one for each component)
counting the sample-category data assignment across all DP nodes.
Number of rows is the number of posterior samples, and number of
columns is the number of data categories.}

\item{\code{comp_dp_counts}}{List of matrices (one for each DP)
counting sample-component assignment (aggregating across data categories).
Number of rows is the number of posterior samples, and number of
columns is the number of components.}

\item{\code{comp_categ_distn}}{List with elements "mean" and "cred.int", containing
matrices with the mean (and lower/upper 95\% credibility interval) distribution
over data categories for each component. Number of rows is the number of
components, and number of columns is the number of data categories. Rows sum to 1.}

\item{\code{comp_dp_distn}}{List with elements "mean" and "cred.int", containing
matrices with the mean (and lower/upper 95\% credibility interval) distribution
over components for each DP. Number of rows is the number of
DPs, and number of columns is the number of components. Rows sum to 1.}
}}

