% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_utils.R
\name{identify_individual_cluster}
\alias{identify_individual_cluster}
\title{Identify individual clusters from ccc.}
\usage{
identify_individual_cluster(main.index, ccc, cdc, disregard)
}
\arguments{
\item{main.index}{the index of cluster in a given ccc}

\item{ccc}{a list clust_categ_counts matrices}

\item{cdc}{a list clust_dp_counts matrices}

\item{disregard}{index of parent nodes to be excluded}
}
\description{
Identify individual clusters from ccc.
}
