% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_components.R
\name{interpret_components}
\alias{interpret_components}
\title{Extract components and exposures from multiple posterior sample chains.}
\usage{
interpret_components(
  multi.chains.retval,
  confident.prop = 0.9,
  noise.prop = 0.5,
  verbose = T
)
}
\arguments{
\item{multi.chains.retval}{A list of objects returned from \code{\link[mSigHdp]{CombineChainsAndExtractSigs}}.}

\item{confident.prop}{Components found in >= \code{confident.prop} of
posterior samples are high confidence components.}

\item{noise.prop}{Components found in < \code{noise.prop} of posterior samples are considered noise components.}

\item{verbose}{if TRUE, generate progress messages.}
}
\value{
In the information that follows, a "component" is
 the union of multiple raw clusters of mutations (in the case of mutational
 signature analysis). Invisibly, a list with the following elements: \describe{

\item{high_confident_components}{A data frame containing the
   components found in >= \code{confident.prop} of posterior samples. Each column is
   a component; in the case of mutational signatures the rows are mutation types.}

\item{high_confident_components_post_number}{A data frame
  in which the first column contains the index of a column in \code{high_confident_components}
 and the second column contains the number of posterior samples that
 contributed to that component.}

\item{high_confident_components_cdc}{A matrix in which each row
 corresponds to one of the Dirichlet processes, and each column
 corresponds to one component in \code{high_confident_components}.
 In the case of mutational signature analysis, most of the columns
 correspond to an input biological sample (e.g. individual tumor).}

\item{moderate_components}{Analogous to \code{high_confident_compents} except for
 components with constituent raw clusters found in >= \code{noise.prop} and
  < \code{confident.prop} posterior samples.}

\item{moderate_components_post_number}{Analogous to \code{high_confident_components_post_number}.}

\item{moderate_components_cdc}{Analogous to \code{high_confident_components_cdc}.}

\item{noise_components}{Analogous to \code{high_confident_compents} except for
 components with constituent raw clusters found in  < \code{noise.prop} posterior samples.}

\item{noise_components_post_number}{Analogous to \code{high_confident_components_post_number}.}

\item{noise_components_cdc}{Analogous to \code{high_confident_components_cdc}.}

\item{extracted.retval}{A list of objects returned from \code{\link[mSigHdp]{CombineChainsAndExtractSigs}}.}

}
}
\description{
This function returns components in which we have high confidence, moderate confidence, and little confidence
(lkely noise).
}
