% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp.R
\docType{data}
\name{lusc_count}
\alias{lusc_count}
\title{Lung squamous cell carcinoma mutation data}
\format{A matrix of mutation counts with 100 rows (one per cancer sample)
 and 96 columns (one per mutation category)}
\usage{
lusc_count
}
\description{
Mutation count data from SomaticCancerAlterations package.
Categories are the 96 base substitution types defined by local trinucleotide
content, and the samples are 100 lung squamous cell carcinomas.
Data is derived from TCGA exome-sequencing studies.
}
\keyword{datasets}
