% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mo_utils.R
\name{mo_plot_comp_size}
\alias{mo_plot_comp_size}
\title{Plot size for each signature for all posterior samples}
\usage{
mo_plot_comp_size(
  retval,
  xlab = "Component",
  ylab = "Number of data items",
  ...
)
}
\arguments{
\item{retval}{an object return from \code{\link{extract_ccc_cdc_from_hdp}}}

\item{xlab}{Horizontal axis label}

\item{ylab}{Vertical axis label}

\item{...}{Other arguments to plot}
}
\description{
Plot size for each signature for all posterior samples
}
