% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp.R
\docType{data}
\name{mut_count}
\alias{mut_count}
\title{Cancer mutation count data}
\format{A matrix of mutation counts with 300 rows (one per cancer sample)
 and 96 columns (one per mutation category)}
\usage{
mut_count
}
\description{
Mutation count data from SomaticCancerAlterations package.
Categories are the 96 base substitution types defined by local trinucleotide
content, and the samples include 100 lung adenocarcinomas, 100 ovarian serous
carcinomas, and 100 skin cutaneous melanomas.
Data is derived from TCGA exome-sequencing studies.
}
\keyword{datasets}
