% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics-output.R
\name{numcomp}
\alias{numcomp}
\title{Get number of extracted components}
\usage{
numcomp(x)
}
\arguments{
\item{x}{hdpSampleChain or hdpSampleMulti}
}
\value{
number of components
}
\description{
Get number of extracted components
}

