% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{plot_chain_hdpsig_exp}
\alias{plot_chain_hdpsig_exp}
\title{Plot hdp signature exposure on each chain}
\usage{
plot_chain_hdpsig_exp(hdpsample, chains, legend = TRUE)
}
\arguments{
\item{hdpsample}{A hdpSampleChain or hdpSampleMulti object including output
from \code{\link{hdp_extract_components}}}

\item{chains}{A hdpSampleChain or hdpSampleMulti object in the list representation}

\item{legend}{Logical - should a legend be included? (default TRUE)}
}
\description{
Plot hdp signature exposure on each chain
}
