% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_components.R
\name{plot_comp_distn}
\alias{plot_comp_distn}
\title{Barplot of the mean distribution over data categories for each component}
\usage{
plot_comp_distn(sample, comp = NULL, cat_names = NULL, grouping = NULL,
  col = "grey70", col_nonsig = NULL, show_group_labels = FALSE,
  cred_int = TRUE, weights = NULL)
}
\arguments{
\item{sample}{A hdpSampleChain or hdpSampleMulti object including output from \code{\link{hdp_extract_components}}}

\item{comp}{(Optional) Number(s) of the component(s) to plot (from 0 to the max component number).
The default is to plot all components.}

\item{cat_names}{(Optional) Data category names to label the horizontal axis}

\item{grouping}{(Optional) A factor indicating data category groups.}

\item{col}{Either a single colour for all data categories, or a vector of
colours for each group (in the same order as the levels of the grouping factor)}

\item{col_nonsig}{(Optional) Colour for any data category whose 95% credibility interval
overlaps with zero (if set, overrides col argument)}

\item{show_group_labels}{Logical - should group labels be added to the top
horizontal axis? (default FALSE) (only works if categories alreayd come in orders)}

\item{cred_int}{Logical - should 95\% credibility intervals be plotted? (default TRUE)}

\item{weights}{(Optional) Weights over the data categories to adjust their
relative contribution (multiplicative)}
}
\description{
Barplot of the mean distribution over data categories for each component
}
\examples{
tcga_example_chain <- hdp_extract_components(tcga_example_chain)

bases <- c("A", "C", "G", "T")
trinuc_context <- paste0(rep(rep(bases, times=4), each=6),
                         rep(c("C", "T"), each=48),
                         rep(bases, times=24))
group_factor <- as.factor(rep(c("C>A", "C>G", "C>T", "T>A", "T>C", "T>G"),
                           each=16))
plot_comp_distn(tcga_example_chain, cat_names=trinuc_context,
                grouping=group_factor, col=RColorBrewer::brewer.pal(6, "Set2"),
                col_nonsig="grey80", show_group_labels=TRUE)

# not run (slow):
# tcga_example_multi <- hdp_extract_components(tcga_example_multi)
# plot_comp_distn(tcga_example_multi, cat_names=trinuc_context,
#                 grouping=group_factor, col=RColorBrewer::brewer.pal(6, "Set2"),
#                 col_nonsig="grey80", show_group_labels=TRUE)
}

