% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_components.R
\name{plot_comp_size}
\alias{plot_comp_size}
\title{Plot number of data items assigned to each component for each posterior sample.}
\usage{
plot_comp_size(chain, legend = TRUE, col_early = "hotpink",
  col_late = "skyblue3", ...)
}
\arguments{
\item{chain}{A hdpSampleChain object including output from \code{\link{hdp_extract_components}}}

\item{legend}{Logical - should a legend be included? (default TRUE)}

\item{col_early}{Color ramp side for early posterior samples}

\item{col_late}{Color ramp side for late posterior samples}

\item{...}{Other arguments to plot}
}
\description{
Plot number of data items assigned to each component for each posterior sample.
}
\examples{
tcga_example_comp <- hdp_extract_components(tcga_example_chain)
plot_comp_size(tcga_example_comp, bty="L")
}
\seealso{
\code{\link{hdp_extract_components}},
 \code{\link{plot_comp_distn_bar}}, \code{\link{plot_dp_comp_exposure}}
}

