% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_plot_functions.R
\name{plot_component_posterior_samples}
\alias{plot_component_posterior_samples}
\title{Plot the distribution of raw clusters highly similar as the component in posterior chains}
\usage{
plot_component_posterior_samples(components, retval)
}
\arguments{
\item{components}{A matrix that containing components with each row corresponding a category and each column
corresponding a component}

\item{retval}{An object return from \code{\link{extract_ccc_cdc_from_hdp}}}
}
\description{
Plot the distribution of raw clusters highly similar as the component in posterior chains
}
