% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_plot_functions.R
\name{plot_component_with_credint}
\alias{plot_component_with_credint}
\title{Plot signatures and their 95\% credible intervals}
\usage{
plot_component_with_credint(
  retval,
  cat_names = NULL,
  col = "grey70",
  cred_int = TRUE,
  weights = NULL,
  group_label_height = 1.05,
  cex.cat = 0.7
)
}
\arguments{
\item{retval}{an object return from \code{\link{extract_ccc_cdc_from_hdp}}.}

\item{cat_names}{names displayed on x-axis, e.g. SBS96 mutation classes}

\item{col}{Either a single colour for all data categories, or a vector of
colours for each group (in the same order as the levels of the grouping factor)}

\item{cred_int}{Logical - should 95\% credibility intervals be plotted? (default TRUE)}

\item{weights}{(Optional) Weights over the data categories to adjust their
relative contribution (multiplicative)}

\item{group_label_height}{Multiplicative factor from top of plot for group label placement}

\item{cex.cat}{Expansion factor for the (optional) cat_names}
}
\description{
Plot signatures and their 95\% credible intervals
}
