% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_chain.R
\name{plot_data_assigned}
\alias{plot_data_assigned}
\title{Plot proportion of data items assigned vs number of clusters}
\usage{
plot_data_assigned(chain, legend = TRUE, col_early = "hotpink",
  col_late = "skyblue3", xlab = "Number of raw clusters",
  ylab = "Cumulative prop. of data assigned", ...)
}
\arguments{
\item{chain}{A hdpSampleChain object}

\item{legend}{Logical - should a legend be included? (default TRUE)}

\item{col_early}{Color ramp side for early posterior samples}

\item{col_late}{Color ramp side for late posterior samples}

\item{xlab}{Horizontal axis label}

\item{ylab}{Vertical axis label}

\item{...}{Other arguments to plot}
}
\description{
Plot the cumulative proportion of data items assigned as the number of raw
clusters increases (raw clusters sorted from largest to smallest).
The horizontal axis extends as far as needed to explain 99.9\% of data
items (on average).
Each posterior sample is plotted separately, using a colour spectrum ranging from
\code{col_early} to \code{col_late}. The average across posterior samples
is plotted in black.
}
\examples{
plot_data_assigned(mut_example_chain, bty="L")
}
\seealso{
\code{\link{hdpSampleChain-class}}, \code{\link{plot_lik}},
 \code{\link{plot_numcluster}}
}

