% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_components.R
\name{plot_dp_comp_exposure}
\alias{plot_dp_comp_exposure}
\title{Plot the mean distribution over components for each specified DP}
\usage{
plot_dp_comp_exposure(chain, dpindices, col, dpnames = NULL,
  main_text = NULL, incl_numdata_plot = TRUE, incl_nonsig = TRUE)
}
\arguments{
\item{chain}{A hdpSampleChain object including output from \code{\link{hdp_extract_components}}}

\item{dpindices}{Indices of DP nodes to plot}

\item{col}{Colours of each component, from 0 to the max number}

\item{dpnames}{(Optional) Names of the DP nodes}

\item{main_text}{(Optional) Text at top of plot}

\item{incl_numdata_plot}{Logical - should an upper barplot indicating the number of
data items per DP be included? (Default TRUE)}

\item{incl_nonsig}{Logical - should components whose credibility intervals include 0
be included (per DP)? (Default TRUE)}
}
\description{
Plot the mean distribution over components for each specified DP
}
\examples{
tcga_example_comp <- hdp_extract_components(tcga_example_chain)
plot_dp_comp_exposure(tcga_example_comp, 4:30,
                      RColorBrewer::brewer.pal(9, "Set3"))
plot_dp_comp_exposure(tcga_example_comp, 4:30,
                      RColorBrewer::brewer.pal(9, "Set3"),
                      incl_numdata_plot=FALSE, incl_nonsig=FALSE)
}
\seealso{
\code{\link{hdp_extract_components}}, \code{\link{plot_comp_size}},
 \code{\link{plot_comp_distn_bar}}
}

