% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_chain.R
\name{plot_lik}
\alias{plot_lik}
\title{Plot likelihood of HDP posterior sampling chain.}
\usage{
plot_lik(chain, start = 1, end = length(lik(chain)), col_lik = "blue",
  col_burn = "red", ...)
}
\arguments{
\item{chain}{A hdpSampleChain object}

\item{start}{The starting iteration to plot from (default 1)}

\item{end}{The final iteration to plot to (default is end of chain)}

\item{col_lik}{Plot colour of likelihood (default blue)}

\item{col_burn}{Plot colour of burnin (default red)}

\item{...}{Other arguments to plot}
}
\description{
Plot likelihood of HDP posterior sampling chain.
}
\examples{
plot_lik(tcga_example_chain, bty="L")
}
\seealso{
\code{\link{hdpSampleChain-class}}, \code{\link{plot_numcluster}},
 \code{\link{plot_data_assigned}}
}

