% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_chain.R
\name{plot_numcluster}
\alias{plot_numcluster}
\title{Plot number of raw clusters in each posterior sample from the chain}
\usage{
plot_numcluster(chain, col = "blue", xlab = "Sample",
  ylab = "Number of raw clusters", ...)
}
\arguments{
\item{chain}{A hdpSampleChain object}

\item{col}{Plot colour (default blue)}

\item{xlab}{Horizontal axis label}

\item{ylab}{Vertical axis label}

\item{...}{Other arguments to plot}
}
\description{
Plot number of raw clusters in each posterior sample from the chain
}
\examples{
plot_numcluster(mut_example_chain, bty="L")
}
\seealso{
\code{\link{hdpSampleChain-class}}, \code{\link{plot_lik}},
 \code{\link{plot_data_assigned}}
}

