% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_plot_functions.R
\name{plot_sig_exposure_for_dp}
\alias{plot_sig_exposure_for_dp}
\title{Plot hdp signature exposure in each sample. This function returns the plot of
top 5 samples with the highest exposure to a component}
\usage{
plot_sig_exposure_for_dp(
  retval,
  hdpsample,
  input.catalog,
  col_comp = NULL,
  incl_numdata_plot = TRUE,
  ylab_numdata = "Number of data items",
  ylab_exp = "Component exposure",
  leg.title = "Component",
  cex.names = 0.6,
  cex.axis = 0.7,
  mar = c(1, 4, 2, 0.5),
  oma = c(1.5, 1.5, 1, 1),
  ...
)
}
\arguments{
\item{retval}{an object return from \code{\link{extract_ccc_cdc_from_hdp}}}

\item{hdpsample}{A hdpSampleChain or hdpSampleMulti object including output
from \code{\link{hdp_extract_components}}}

\item{input.catalog}{input catalog for samples}

\item{col_comp}{Colours of each component, from 0 to the max number. If NULL,                          default colors will be used}

\item{incl_numdata_plot}{Logical - should an upper barplot indicating the number of
data items per DP be included? (Default TRUE)}

\item{ylab_numdata}{Vertical axis label for numdata plot}

\item{ylab_exp}{Vertical exis label for exposure plot}

\item{leg.title}{Legend title}

\item{cex.names}{Expansion factor for bar labels (dpnames) in exposure plot}

\item{cex.axis}{Expansion factor for vertical-axis annotation}

\item{mar}{See ?par}

\item{oma}{See ?par}

\item{...}{Other arguments to plot}
}
\description{
Plot hdp signature exposure in each sample. This function returns the plot of
top 5 samples with the highest exposure to a component
}
