% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chain.R
\name{plotchain}
\alias{plotchain}
\alias{plot_lik}
\alias{plot_numcluster}
\alias{plot_data_assigned}
\title{Diagnostic plots for HDP posterior sampling chain}
\usage{
plot_lik(chain, start = 1, end = length(lik(chain)), col_lik = "blue",
  col_burn = "red", xlab = "Iteration", ylab = "Likelihood", ...)

plot_numcluster(chain, col = "blue", xlab = "Sample",
  ylab = "Number of raw clusters", ...)

plot_data_assigned(chain, legend = TRUE, col_early = "hotpink",
  col_late = "skyblue3", xlab = "Number of raw clusters",
  ylab = "Cumulative prop. of data assigned", ...)
}
\arguments{
\item{chain}{A hdpSampleChain object}

\item{start}{The starting iteration to plot from (default 1)}

\item{end}{The final iteration to plot to (default is end of chain)}

\item{col_lik}{Plot colour of likelihood (default blue)}

\item{col_burn}{Plot colour of burnin (default red)}

\item{xlab}{Horizontal axis label}

\item{ylab}{Vertical axis label}

\item{...}{Other arguments to plot}

\item{col}{Plot colour for numcluster (default blue)}

\item{legend}{Logical - should a legend be included? (default TRUE)}

\item{col_early}{Color ramp side for early posterior samples}

\item{col_late}{Color ramp side for late posterior samples}
}
\description{
Diagnostic plots for HDP posterior sampling chain
}
\examples{
plot_lik(mut_example_chain, bty="L")
plot_numcluster(mut_example_chain, bty="L")
plot_data_assigned(mut_example_chain, bty="L")
}
