% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_components.R
\name{plotcomp}
\alias{plotcomp}
\alias{plot_comp_size}
\alias{plot_comp_distn}
\alias{plot_dp_comp_exposure}
\alias{plot_tsne_sigs_tumortype}
\title{Plot extracted components}
\usage{
plot_comp_size(
  hdpsample,
  legend = TRUE,
  col_a = "hotpink",
  col_b = "skyblue3",
  xlab = "Component",
  ylab = "Number of data items",
  ...
)

plot_comp_distn(
  hdpsample,
  comp = NULL,
  cat_names = NULL,
  grouping = NULL,
  col = "grey70",
  col_nonsig = NULL,
  show_group_labels = FALSE,
  cred_int = TRUE,
  weights = NULL,
  plot_title = NULL,
  group_label_height = 1.05,
  cex.cat = 0.7,
  ...
)

plot_dp_comp_exposure(
  hdpsample,
  input.catalog,
  col_comp,
  dpnames = NULL,
  main_text = NULL,
  incl_numdata_plot = TRUE,
  incl_nonsig = TRUE,
  incl_comp0 = TRUE,
  ylab_numdata = "Number of data items",
  ylab_exp = "Component exposure",
  leg.title = "Component",
  cex.names = 0.6,
  cex.axis = 0.7,
  mar = c(1, 4, 2, 0.5),
  oma = c(1.5, 1.5, 1, 1),
  ...
)

plot_tsne_sigs_tumortype(exposure, colorcategory = NULL)
}
\arguments{
\item{hdpsample}{A hdpSampleChain or hdpSampleMulti object including output
from \code{\link{hdp_extract_components}}}

\item{legend}{Logical - should a legend be included? (default TRUE)}

\item{col_a}{Color ramp side for early posterior samples (if hdpSampleChain) or first chain (if hdpSampleMulti)}

\item{col_b}{Color ramp side for late posterior samples (if hdpSampleChain) or last chain (if hdpSampleMulti)}

\item{xlab}{Horizontal axis label}

\item{ylab}{Vertical axis label}

\item{...}{Other arguments to plot}

\item{comp}{(Optional) Number(s) of the component(s) to plot (from 0 to the max component number).
The default is to plot all components.}

\item{cat_names}{(Optional) Data category names to label the horizontal axis}

\item{grouping}{(Optional) A factor indicating data category groups.}

\item{col}{Either a single colour for all data categories, or a vector of
colours for each group (in the same order as the levels of the grouping factor)}

\item{col_nonsig}{(Optional) Colour for any data category whose 95\% credibility interval
overlaps with zero (if set, overrides col argument)}

\item{show_group_labels}{Logical - should group labels be added to the top
horizontal axis? (default FALSE) (only works if categories alreayd come in orders)}

\item{cred_int}{Logical - should 95\% credibility intervals be plotted? (default TRUE)}

\item{weights}{(Optional) Weights over the data categories to adjust their
relative contribution (multiplicative)}

\item{plot_title}{(Optional) Character vector of custom plot titles (one for each component plotted)}

\item{group_label_height}{Multiplicative factor from top of plot for group label placement}

\item{cex.cat}{Expansion factor for the (optional) cat_names}

\item{input.catalog}{input catalog for samples}

\item{col_comp}{Colours of each component, from 0 to the max number}

\item{dpnames}{(Optional) Names of the DP nodes}

\item{main_text}{(Optional) Text at top of plot}

\item{incl_numdata_plot}{Logical - should an upper barplot indicating the number of
data items per DP be included? (Default TRUE)}

\item{incl_nonsig}{Logical - should components whose credibility intervals include 0
be included (per DP)? (Default TRUE)}

\item{incl_comp0}{Logical - should component zero be plotted? (Default TRUE)}

\item{ylab_numdata}{Vertical axis label for numdata plot}

\item{ylab_exp}{Vertical exis label for exposure plot}

\item{leg.title}{Legend title}

\item{cex.names}{Expansion factor for bar labels (dpnames) in exposure plot}

\item{cex.axis}{Expansion factor for vertical-axis annotation}

\item{mar}{See ?par}

\item{oma}{See ?par}

\item{exposure}{An exposure matrix}

\item{colorcategory}{A list of category of samples for labelling tsne plot.
Length is equal as number of samples}
}
\description{
Plot extracted components

Plot hdp signature exposure in each sample

A tsne plot for signatures and tumors
}
