% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predictive}
\alias{predictive}
\title{predictive_marginal}
\usage{
predictive(
  idx_group,
  grid,
  fit,
  priorMean,
  priorA,
  priorB,
  priorLambda,
  burnin = 0
)
}
\value{
[matrix] of size \code{n x length(grid)} containing the quantiles of level \code{0.025,0.5,0.975}.
}
\description{
predictive_marginal
}
