% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predictive_all_groups}
\alias{predictive_all_groups}
\title{predictive_all_groups}
\usage{
predictive_all_groups(
  grid,
  fit,
  priorMean,
  priorA,
  priorB,
  priorLambda,
  burnin = 0
)
}
\value{
[list] of length \code{d} where each element is the return object of \code{\link{predictive}}.
}
\description{
predictive_all_groups
}
