% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics-output.R
\name{prop.ex}
\alias{prop.ex}
\title{Get proportion of dataset explained (on average)}
\usage{
prop.ex(x)
}
\arguments{
\item{x}{hdpSampleChain or hdpSampleMulti}
}
\value{
number of components
}
\description{
Get proportion of dataset explained (on average)
}
