% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdp.R
\docType{data}
\name{tcga_example_chain}
\alias{tcga_example_chain}
\title{Posterior sampling chain from TCGA somatic mutation data}
\format{A hdpSampleChain object with 50 posterior samples}
\usage{
tcga_example_chain
}
\description{
Describe here
}
\keyword{datasets}

