% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdp.R
\docType{data}
\name{tcga_example_multi}
\alias{tcga_example_multi}
\title{Multiple posterior sampling chains with TCGA somatic mutation data}
\format{A hdpSampleMulti object with 250 posterior samples, 50 from each chain}
\usage{
tcga_example_multi
}
\description{
Five independent HDP posterior sampling chains with data from SomaticCancerAlterations package.
Categories are the 96 base substitution types defined by local trinucleotide
content, and the samples include 291 glioblastoma
multiforme (gbm) samples, and 293 kidney chromophobe cancer (kirc) samples.
Data was derived from exome-sequencing studies.
Each sample was assigned to a unique child DP node, with one parent DP node for
gbm, one parent node for kirc samples, and one grandparent DP node at the top level.
Each of the five chains were initialised with 8 clusters, then run through 3000 burn-in iterations before
collecting 50 posterior samples with 50 iterations between each.
}
\keyword{datasets}

