% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-generics-output.R
\name{comp_categ_distn}
\alias{comp_categ_distn}
\title{Get mean distribution over data categories for each component}
\usage{
comp_categ_distn(x)
}
\arguments{
\item{x}{hdpSampleChain or hdpSampleMulti}
}
\value{
List with elements "mean" and "cred.int", containing
 matrices with the mean (and lower/upper 95\% credibility interval) distribution
 over data categories for each component. Number of rows is the number of
 components, and number of columns is the number of data categories. Rows sum to 1.
}
\description{
Get mean distribution over data categories for each component
}
