% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaa-generics.R
\docType{methods}
\name{comp_dp_distn}
\alias{comp_dp_distn}
\alias{comp_dp_distn,hdpSampleChain-method}
\alias{comp_dp_distn,hdpSampleMulti-method}
\title{Get mean distribution over components for each DP}
\usage{
comp_dp_distn(x)

\S4method{comp_dp_distn}{hdpSampleChain}(x)

\S4method{comp_dp_distn}{hdpSampleMulti}(x)
}
\arguments{
\item{x}{hdpSampleChain or hdpSampleMulti}
}
\value{
List with elements "mean" and "cred.int", containing
 matrices with the mean (and lower/upper 95\% credibility interval) distribution
 over components for each DP. Number of rows is the number of
 DPs, and number of columns is the number of components. Rows sum to 1.
}
\description{
Get mean distribution over components for each DP
}
\section{Methods (by class)}{
\itemize{
\item \code{hdpSampleChain}: 

\item \code{hdpSampleMulti}: 
}}

