% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dp_activate.R
\name{dp_activate}
\alias{dp_activate}
\title{Activate DP nodes}
\usage{
dp_activate(hdp, dpindex, initcc)
}
\arguments{
\item{hdp}{A hdpState object}

\item{dpindex}{Indices of the DPs to activate (include all parent DPs)}

\item{initcc}{Number of data clusters to start with
(every data item is randomly assigned to a cluster to start with)}
}
\value{
A hdpState object with activated DPs and an initial random cluster allocation for each data item. See \code{\link{hdpState-class}}
}
\description{
Specifiy the number of starting clusters, and activate the DP nodes to be
included in the posterior sampling process (\code{\link{hdp_posterior}}).
When initialised, the DP nodes are 'heldout' (not available for posterior sampling).
}
\details{
Note that this step can be slow and memory-intensive for very large datasets.
}
\examples{
my_hdp <- hdp_init(ppindex=0, cpindex=1, hh=rep(1, 6), alphaa=rep(1, 3), alphab=rep(2, 3))
my_hdp <- hdp_adddp(my_hdp, 2, 1, 2)
my_hdp <- hdp_adddp(my_hdp, 10, c(rep(2, 5), rep(3, 5)), 3)
my_hdp <- hdp_setdata(my_hdp, 4:13, example_data_hdp)
# active all DPs and start with two data clusters
my_hdp <- dp_activate(my_hdp, 1:13, 2)
}
\seealso{
\code{\link{hdp_init}}, \code{\link{hdp_addconparam}}, \code{\link{hdp_adddp}},
 \code{\link{hdp_setdata}}, \code{\link{hdp_posterior}}
}

