% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dp_freeze.R
\name{dp_freeze}
\alias{dp_freeze}
\title{Freeze DP nodes}
\usage{
dp_freeze(hdp, dpindex)
}
\arguments{
\item{hdp}{A hdpState object}

\item{dpindex}{Indices of the DPs to freeze}
}
\value{
A hdpState object with the specified DP nodes frozen. See \code{\link{hdpState-class}}
}
\description{
Freezes previously active DP nodes. A frozen DP node is not included in posterior sampling,
but its statistics \emph{are} considered in the sampling of other active DPs.
This is useful for conditioning on a previous dataset. First, set up a HDP
for one dataset, run the posterior sampling chain, and then freeze all old nodes
(except the top DP). Add new DP nodes with new data and run a
second posterior sampling chain over the new nodes (\emph{given} the information in the frozen nodes).
}
\seealso{
\code{\link{hdp_init}}, \code{\link{hdp_addconparam}}, \code{\link{hdp_adddp}},
 \code{\link{hdp_setdata}}, \code{\link{dp_activate}}, \code{\link{hdp_posterior}}
}

