% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sigs_from_clusters.R
\name{extract_sigs_from_clusters}
\alias{extract_sigs_from_clusters}
\title{Combine "raw clusters" of mutations in "components" (clusters of mutations generated by one mutational process).}
\usage{
extract_sigs_from_clusters(x, cos.merge = 0.9, hc.cutoff = 0.12)
}
\arguments{
\item{x}{An \code{\link{hdpSampleChain-class}} or \
code{\link{hdpSampleMulti-class}} object}

\item{cos.merge}{Merge raw cluster that have cosine similarity above this
threshold.}

\item{hc.cutoff}{the height to cut hierarchical clustering tree}
}
\value{
A list with the elements \describe{
clustered.spectrum = spectrum.df,
\item{stats.post.samples}{}
\item{spectrum.cdc}{}
\item{each.chain.noise.cd}{A list of raw clusters that }
\item{each.chain.noise.spectrum}{noise from each chain}
\item{multi.chains}{An \code{\link{hdpSampleChain-class}}
   or \code{\link{hdpSampleMulti-class}}
   object updated with component information
}
\item{nsamp}{The total number of posterior samples.}
}
}
\description{
Combine "raw clusters" of mutations in "components" (clusters of mutations generated by one mutational process).
}
\seealso{
\code{\link{hdp_posterior}}, \code{\link{hdp_multi_chain}},
 \code{\link{plot_comp_size}}, \code{\link{plot_comp_distn}},
 \code{\link{plot_dp_comp_exposure}}
}
