% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sigs_from_clusters.R
\name{extract_sigs_from_clusters}
\alias{extract_sigs_from_clusters}
\title{This function extract signatures from raw clusters discovered by hdp.
This returns clusters with the number of posterior samples that they belong}
\usage{
extract_sigs_from_clusters(x, cos.merge = 0.9, hc.cutoff = 0.12)
}
\arguments{
\item{x}{\code{hdpSampleChain} or \code{hdpSampleMulti} object}

\item{cos.merge}{Merge components with cosine similarity above this threshold (default 0.90).}

\item{hc.cutoff}{the height to cut hierarchcial clustering tree}
}
\value{
A hdpSampleChain or hdpSampleMulti object updated with component information
}
\description{
ccc_0 and cdc_0 are for diagnostic checking and exposure summary
}
\seealso{
\code{\link{hdp_posterior}}, \code{\link{hdp_multi_chain}},
 \code{\link{plot_comp_size}}, \code{\link{plot_comp_distn}},
 \code{\link{plot_dp_comp_exposure}}
}
