% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sigs_from_clusters.R
\name{extract_sigs_from_clusters}
\alias{extract_sigs_from_clusters}
\title{This function extract signatures from raw clusters discovered by hdp.
This returns clusters with high confidence -extracted by more than 90% posterior samples;
                           moderate confidence - extracted less than 90% but more than number of post.samples on one chain
                           noise - less than number of post.samples on one chain
and they are found in how many post.samples in 'stats'}
\usage{
extract_sigs_from_clusters(
  x,
  cos.merge = 0.9,
  confident.prop = 0.9,
  noise.prop = 0.5
)
}
\arguments{
\item{x}{\code{hdpSampleChain} or \code{hdpSampleMulti} object}

\item{cos.merge}{Merge components with cosine similarity above this threshold (default 0.90).}

\item{confident.prop}{a numeric between 0.1 and 1. A cluster with more than                           confident.prop of total posterior samples will be determined as high                    confident signature}

\item{noise.prop}{a numeric between 0.1 and 1. A cluster with less than                    noise.prop of total posterior samples will be determined as noise signature}
}
\value{
A hdpSampleChain or hdpSampleMulti object updated with component information
}
\description{
ccc_0 and cdc_0 are for diagnostic checking and exposure summary
}
\seealso{
\code{\link{hdp_posterior}}, \code{\link{hdp_multi_chain}},
 \code{\link{plot_comp_size}}, \code{\link{plot_comp_distn}},
 \code{\link{plot_dp_comp_exposure}}
}
