% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_adddp.R
\name{hdp_adddp}
\alias{hdp_adddp}
\title{Add DPs to a hdpState object}
\usage{
hdp_adddp(hdp, numdp, ppindex, cpindex)
}
\arguments{
\item{hdp}{A hdpState object}

\item{numdp}{The number of DPs to add}

\item{ppindex}{Index (or indices) of the parental process(es) for the new DPs.}

\item{cpindex}{Index (or indices) of the concentration parameters for the new DPs.}
}
\value{
A hdpState object with the updated HDP structure. See \code{\link{hdpState-class}}
}
\description{
Add DP nodes to a hdpState object and specify each parent relationship
and concentration parameter. Concentration parameters can be added to a
hdpState object with \code{\link{hdp_addconparam}}. Data is assigned via \code{\link{hdp_setdata}}.
When initialised, the DP nodes are 'heldout' (not available for posterior sampling)
and will need to be activated (see \code{\link{dp_activate}}). Finally, the posterior
sampling process (a Gibbs sampler) is run via \code{\link{hdp_posterior}}.
}
\examples{
my_hdp <- hdp_init(ppindex=0, cpindex=1, hh=rep(1, 6), alphaa=rep(1, 3), alphab=rep(2, 3))
# add two more DPs with parent '1' and concentration parameter '2'
my_hdp <- hdp_adddp(my_hdp, 2, 1, 2)
my_hdp
hdp_example <- hdp_init(c(0, 1, 1), c(1, 2, 2), rep(1, 6), rep(2, 2), rep(0.5, 2))
# add six more DPs, three with parent '2', three with parent '3',
# and all with concentration parameter '2'
hdp_example <- hdp_adddp(hdp_example, 6, c(2, 2, 2, 3, 3, 3), 2)
hdp_example
}
\seealso{
\code{\link{hdp_init}}, \code{\link{hdp_setdata}},
 \code{\link{dp_activate}}, \code{\link{hdp_posterior}}
}
