% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaa-generics.R
\docType{methods}
\name{hdp_comp_settings}
\alias{hdp_comp_settings}
\alias{hdp_comp_settings,hdpSampleChain-method}
\alias{hdp_comp_settings,hdpSampleMulti-method}
\title{Get settings of component extraction from hdpSampleChain or hdpSampleMulti}
\usage{
hdp_comp_settings(x)

\S4method{hdp_comp_settings}{hdpSampleChain}(x)

\S4method{hdp_comp_settings}{hdpSampleMulti}(x)
}
\arguments{
\item{x}{hdpSampleChain or hdpSampleMulti}
}
\value{
list of settings from hdp_extract_components
}
\description{
Get settings of component extraction from hdpSampleChain or hdpSampleMulti
}
\section{Methods (by class)}{
\itemize{
\item \code{hdpSampleChain}: 

\item \code{hdpSampleMulti}: 
}}

