% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdp_extract_components.R
\name{hdp_extract_components}
\alias{hdp_extract_components}
\title{Extract major components from the raw clusters}
\usage{
hdp_extract_components(chains, cos.merge = 0.9, min.sample = 1)
}
\arguments{
\item{chains}{hdpSampleChain or hdpSampleMulti object}

\item{cos.merge}{Merge components with cosine similarity above this threshold (default 0.90)}

\item{min.sample}{Components must have significant exposure in at least this many samples (i.e. those DP nodes with data assigned) (default 1)}
}
\value{
A hdpSampleChain or hdpSampleMulti object updated with component information
}
\description{
If prior components included via \code{\link{hdp_prior_init}} are
preserved by \code{hdp_extract_components}, they are prefixed with "P".
Any new components in this case are prefixed with "N".
}
\examples{
hdp_extract_components(mut_example_chain)
}
\seealso{
\code{\link{hdp_posterior}}, \code{\link{hdp_multi_chain}},
 \code{\link{plot_comp_size}}, \code{\link{plot_comp_distn}},
 \code{\link{plot_dp_comp_exposure}}
}
