% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaa-generics-components.R
\docType{methods}
\name{hdp_extract_components}
\alias{hdp_extract_components}
\alias{hdp_extract_components,hdpSampleChain-method}
\alias{hdp_extract_components,hdpSampleMulti-method}
\title{Extract major components from the raw clusters}
\usage{
hdp_extract_components(x, cos.merge = 0.9, prop.ex = 0.97, redo = TRUE)

\S4method{hdp_extract_components}{hdpSampleChain}(x, cos.merge, prop.ex)

\S4method{hdp_extract_components}{hdpSampleMulti}(x, cos.merge, redo)
}
\arguments{
\item{x}{hdpSampleChain or hdpSampleMulti object}

\item{cos.merge}{Merge components with cosine similarity above this threshold (default 0.90)}

\item{prop.ex}{The proportion of data items explained by the components (default 0.97).
Only used for hdpSampleChain as this can be calculated automatically for hdpMultiChain objects.}

\item{redo}{Logical. If true - constituent chains with previously calculated components
will be re-calculated. Only used for hdpSampleMulti.}
}
\value{
A hdpSampleChain or hdpSampleMulti object updated with component information
}
\description{
Extract major components from the raw clusters
}
\section{Methods (by class)}{
\itemize{
\item \code{hdpSampleChain}: 

\item \code{hdpSampleMulti}: 
}}
\examples{
hdp_extract_components(tcga_example_chain)
}

