% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hdp_posterior}
\alias{hdp_posterior}
\title{Posterior sampling across activated DPs.}
\usage{
hdp_posterior(hdp, numburnin, numsample, numspace, doconparam, dolik = 1,
  dodebug = 0)
}
\arguments{
\item{hdp}{A HDP list object.}

\item{numburnin}{The number of burnin iterations.}

\item{numsample}{The number of posterior samples to collect.}

\item{numspace}{The number of iterations to leave between collected samples.}

\item{doconparam}{The number of iterations of concentration parameter sampling to perform between iterations (?)}

\item{dolik}{Leave as default (1)}

\item{dodebug}{Verbosity of debugging statements. 0 is least verbose, 4 is most verbose. 0 Highly recommended.}
}
\value{
A list with seven elements: 'hdp' - the HDP list object at the end of the sampling process; 'lik' - a numeric vector of the HDP likelihood over the entire sampling process; 'numclass' - a numeric vector of the total number of classes/clusters in each posterior sample collected; 'classqq' - a list of matrices summarising the total number of data points of each category in each overall class/cluster at each posterior sample collected; 'classnd' - a list of matrices summarising the total number of data points in each DP in each overall class/cluster at each posterior sample collected; 'alpha'; and 'beta'.
}
\description{
Posterior sampling across activated DPs.
}

