% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hdp_setdata.R
\name{hdp_setdata}
\alias{hdp_setdata}
\title{Assign data to DP nodes in a hdpState object}
\usage{
hdp_setdata(hdp, dpindex, data)
}
\arguments{
\item{hdp}{A hdpState object}

\item{dpindex}{Indices of the DPs to assign data to (in same order as rows of \code{data})}

\item{data}{A \code{data.frame} or \code{matrix} of counts with one row for every sample
(same order as \code{dpindex}) and one column for every data category.}
}
\value{
A hdpState object updated with the new data. See \code{\link{hdpState-class}}
}
\description{
Assign data to 'heldout' (state is 0) DP nodes in a hdpState object.
'Heldout' DPs are not available for posterior sampling,
and will need to be activated (see \code{\link{dp_activate}}). The posterior
sampling process (a Gibbs sampler) is run via \code{\link{hdp_posterior}}.
}
\examples{
example_data_hdp
my_hdp <- hdp_init(ppindex=0, cpindex=1, hh=rep(1, 6), alphaa=rep(1, 3), alphab=rep(2, 3))
my_hdp <- hdp_adddp(my_hdp, 2, 1, 2)
my_hdp <- hdp_adddp(my_hdp, 10, c(rep(2, 5), rep(3, 5)), 3)
my_hdp <- hdp_setdata(my_hdp, 4:13, example_data_hdp)
dp(my_hdp)
}
\seealso{
\code{\link{hdp_init}}, \code{\link{hdp_adddp}},
 \code{\link{dp_activate}}, \code{\link{hdp_posterior}}
}

