% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aaa-classes.R, R/aaa-generics.R
\docType{class}
\name{hdpSampleMulti-class}
\alias{as.list,hdpSampleMulti-method}
\alias{chains}
\alias{chains,hdpSampleMulti-method}
\alias{hdpSampleMulti-class}
\title{hdpSampleMulti class for multiple independent hdpSampleChain objects for the same HDP}
\usage{
\S4method{as.list}{hdpSampleMulti}(x, ...)

\S4method{chains}{hdpSampleMulti}(x, ...)
}
\arguments{
\item{x}{Object of class hdpSampleMulti}

\item{...}{unused}
}
\description{
hdpSampleMulti class for multiple independent hdpSampleChain objects for the same HDP
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.list}: Convert to list class

\item \code{chains}: Get list of hdpSampleChain objects
}}
\section{Slots}{

\describe{
\item{\code{chains}}{List of hdpSampleChain objects storing multiple independent runs of the posterior sampling chain for the same data and HDP struct}

\item{\code{comp_settings}}{Settings used to
consolidate raw clusters (variable number across posterior samples) into
global components (constant number across posterior samples):
  cos.merge (merge components within cosine similarity)}

\item{\code{comp_categ_counts}}{List of matrices (one for each component)
counting the sample-category data assignment across all DP nodes.
Number of rows is the number of posterior samples, and number of
columns is the number of data categories.}

\item{\code{comp_dp_counts}}{List of matrices (one for each DP)
counting sample-component assignment (aggregating across data categories).
Number of rows is the number of posterior samples, and number of
columns is the number of components.}

\item{\code{comp_dp_weights}}{List of matrices (one for each DP)
with the weights of each component per sample. Number of rows is the number of
posterior samples, and number of columns is the number of components.}

\item{\code{comp_categ_distn}}{List with elements "mean" and "cred.int", containing
matrices with the mean (and lower/upper 95% credibility interval) distribution
over data categories for each component. Number of rows is the number of
components, and number of columns is the number of data categories. Rows sum to 1.}

\item{\code{comp_dp_distn}}{List with elements "mean" and "cred.int", containing
matrices with the mean (and lower/upper 95% credibility interval) distribution
over components for each DP. Number of rows is the number of
DPs, and number of columns is the number of components. Rows sum to 1.}

\item{\code{prop.ex}}{(Average) proportion of dataset explained by the components consistently
found in all chains.}
}}

