% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa-classes-input.R, R/aaa-generics-input.R
\docType{class}
\name{hdpState-class}
\alias{activating_seed}
\alias{activating_seed,hdpState-method}
\alias{as.list,hdpState-method}
\alias{base}
\alias{base,hdpState-method}
\alias{base_params}
\alias{base_params,hdpState-method}
\alias{conparam}
\alias{conparam,hdpState-method}
\alias{cpindex}
\alias{cpindex,hdpState-method}
\alias{dp}
\alias{dp,hdpState-method}
\alias{dpstate}
\alias{dpstate,hdpState-method}
\alias{hdpState-class}
\alias{numcateg}
\alias{numcateg,hdpState-method}
\alias{numconparam}
\alias{numconparam,hdpState-method}
\alias{numdp}
\alias{numdp,hdpState-method}
\alias{ppindex}
\alias{ppindex,hdpState-method}
\alias{pseudoDP}
\alias{pseudoDP,hdpState-method}
\title{hdpState class for a Hierarchical Dirichlet Process in one state}
\usage{
\S4method{as.list}{hdpState}(x, ...)

\S4method{numdp}{hdpState}(x, ...)

\S4method{numconparam}{hdpState}(x, ...)

\S4method{base}{hdpState}(x, ...)

\S4method{conparam}{hdpState}(x, ...)

\S4method{dp}{hdpState}(x, ...)

\S4method{dpstate}{hdpState}(x, ...)

\S4method{ppindex}{hdpState}(x, ...)

\S4method{cpindex}{hdpState}(x, ...)

\S4method{numcateg}{hdpState}(x, ...)

\S4method{base_params}{hdpState}(x, ...)

\S4method{activating_seed}{hdpState}(x, ...)

\S4method{pseudoDP}{hdpState}(x, ...)
}
\arguments{
\item{x}{Object of class hdpState}

\item{...}{unused}
}
\description{
hdpState class for a Hierarchical Dirichlet Process in one state
}
\section{Methods (by generic)}{
\itemize{
\item \code{as.list}: Convert to list class

\item \code{numdp}: Get number of DPs

\item \code{numconparam}: Get number of concentration parameters

\item \code{base}: Get base distribution

\item \code{conparam}: Get list of concentration parameters

\item \code{dp}: Get list of DP nodes

\item \code{dpstate}: Get state of every DP

\item \code{ppindex}: Get parent process index of every DP

\item \code{cpindex}: Get concentration parameter index of every DP

\item \code{numcateg}: Get number of data categories

\item \code{base_params}: Get parameters of the base Dirichlet distribution (like psuedocounts across categories).

\item \code{activating_seed}: Get seed used to initialse clustering

\item \code{pseudoDP}: Get index of frozen pseudo-data DP nodes for prior info
(only if initialised via hdp_prior_init)
}}
\section{Slots}{

\describe{
\item{\code{numdp}}{number of DP nodes in the hierarchical Dirichlet Process}

\item{\code{numconparam}}{number of concentration parameters}

\item{\code{base}}{base distribution (hdpBase object)}

\item{\code{conparam}}{concentration parameters (list of hdpConparam objects)}

\item{\code{dp}}{DP nodes (list of hdpDP objects)}

\item{\code{dpstate}}{state of DP nodes for posterior sampling process: active (2), frozen (1), or heldout (0)}

\item{\code{ppindex}}{parent node index for each DP}

\item{\code{cpindex}}{concentration parameter index for each DP}

\item{\code{ttindex}}{DP index of those sharing a concentration parameter}

\item{\code{initcc}}{number of initial clusters}

\item{\code{seed_activate}}{random seed used to initiate cluster membership}

\item{\code{pseudoDP}}{(Optional) index of pseudodata nodes (only if initialised via hdp_prior_init)}
}}

