% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret_components.R
\name{interpret_components}
\alias{interpret_components}
\title{Extract components and exposures from multiple posterior sample chains
This function returns components with high confidence, moderate confidence and possible noise}
\usage{
interpret_components(
  multi.chains.retval,
  confident.prop = 0.9,
  noise.prop = 0.5,
  verbose = T
)
}
\arguments{
\item{multi.chains.retval}{A list of objects returned from \code{\link[mSigHdp]{CombineChainsAndExtractSigs}}.}

\item{confident.prop}{components with at least \code{confident.prop} of posterior samples are high confident signatures}

\item{noise.prop}{components with less than \code{noise.prop} of posterior samples are noise signatures}

\item{verbose}{if TRUE, return messages}
}
\value{
Invisibly, a list with the following elements:\describe{

\item{high_confident_components}{The components found in more than \code{confident.prop} of posterior samples.
                               The components that are profiles as a matrix;rows are catalogy types(features), columns are index.}
\item{high_confident_components_post_number}{number of posterior samples corresponding
                                           to each cluster in \code{high_confident_components}.}
\item{high_confident_components_cdc}{categ_dp_counts matrix;rows are dps, columns correspond
                                   to each cluster in \code{high_confident_components}}
\item{moderate_components}{The components found in more than \code{noise.prop} but less than \code{confident.prop} of posterior samples.
                         The components that are profiles as a matrix; rows are catalogy types(features),columns are index.}
\item{moderate_components_post_number}{number of posterior samples corresponding to each cluster in
                                          \code{moderate_components}.}
\item{moderate_components_cdc}{categ_dp_counts matrix;rows are dps, columns correspond to each cluster
                                   in \code{moderate_components}}
\item{noise_components}{The components found in less than \code{noise.prop} of posterior samples.
                     The components that are profiles as a matrix;rows are catalogy types(features), columns are index.}
\item{noise_components_post_number}{number of posterior samples corresponding to each cluster in
                                          \code{noise_components}.}
\item{noise_components_cdc}{categ_dp_counts matrix;rows are dps, columns correspond to each cluster
                                   in \code{noise_components}}
\item{extracted.retval}{A list of objects returned from \code{\link[mSigHdp]{CombineChainsAndExtractSigs}}}

}
}
\description{
Extract components and exposures from multiple posterior sample chains
This function returns components with high confidence, moderate confidence and possible noise
}
