% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_components.R
\name{plot_comp_size}
\alias{plot_comp_size}
\title{Plot number of data items assigned to each component for each posterior sample.}
\usage{
plot_comp_size(sample, legend = TRUE, col_a = "hotpink",
  col_b = "skyblue3", ...)
}
\arguments{
\item{sample}{A hdpSampleChain or hdpSampleMulti object including
output from \code{\link{hdp_extract_components}}}

\item{legend}{Logical - should a legend be included? (default TRUE)}

\item{col_a}{Color ramp side for early posterior samples (if hdpSampleChain) or first chain (if hdpSampleMulti)}

\item{col_b}{Color ramp side for late posterior samples (if hdpSampleChain) or last chain (if hdpSampleMulti)}

\item{...}{Other arguments to plot}
}
\description{
Plot number of data items assigned to each component for each posterior sample.
}
\examples{
tcga_example_chain <- hdp_extract_components(tcga_example_chain)
plot_comp_size(tcga_example_chain, bty="L")

# not run (slow):
# tcga_example_multi <- hdp_extract_components(tcga_example_multi)
# plot_comp_size(tcga_example_multi, bty="L")
}

